[lldb][NFC] Move discovery of local variables earlier in expr eval
This patch moves the discovery of local variables form
SwiftExpressionParser to SwiftUserExpression. This has two main
benefits:
- In case we're binding generic parameters, we can fail earlier if we
  can't bind some variable.
- Prepares the path to implement expression evaluation with an unlimited
  amount of generic parameters.
