+++ lldb/source/Plugins/ExpressionParser/Swift/SwiftUserExpression.cpp
@@ -10,14 +10,17 @@
 //
 //===----------------------------------------------------------------------===//
 
+#include "swift/AST/ASTContext.h"
+#include "swift/Demangling/Demangler.h"
 #include <stdio.h>
 #if HAVE_SYS_TYPES_H
 #include <sys/types.h>
 #endif
 
+#include "SwiftASTManipulator.h""
 #include "SwiftExpressionParser.h"
-#include "SwiftREPLMaterializer.h"
 #include "SwiftExpressionSourceCode.h"
+#include "SwiftREPLMaterializer.h"
 
 #include "Plugins/LanguageRuntime/Swift/SwiftLanguageRuntime.h"
 #include "lldb/Core/Module.h"
@@ -262,6 +265,184 @@ void SwiftUserExpression::ScanContext(ExecutionContext &exe_ctx, Status &err) {
             info.type.GetTypeName().AsCString());
 }
 
+/// Create a \c VariableInfo record for \c variable if there isn't
+ /// already shadowing inner declaration in \c processed_variables.
+static bool AddVariableInfo(
+    lldb::VariableSP variable_sp, lldb::StackFrameSP &stack_frame_sp,
+    SwiftASTContextForExpressions &ast_context, SwiftLanguageRuntime *runtime,
+    llvm::SmallDenseSet<const char *, 8> &processed_variables,
+    llvm::SmallVectorImpl<SwiftASTManipulator::VariableInfo> &local_variables,
+    lldb::DynamicValueType use_dynamic,
+    lldb::BindGenericTypes bind_generic_types) {
+  LLDB_SCOPED_TIMER();
+
+  StringRef name = variable_sp->GetUnqualifiedName().GetStringRef();
+  const char *name_cstr = name.data();
+  assert(StringRef(name_cstr) == name && "missing null terminator");
+  if (name.empty())
+    return true;
+
+  // To support "guard let self = self" the function argument "self"
+  // is processed (as the special self argument) even if it is
+  // shadowed by a local variable.
+  bool is_self = SwiftLanguageRuntime::IsSelf(*variable_sp);
+  const char *overridden_name = name_cstr;
+  if (is_self)
+    overridden_name = "$__lldb_injected_self";
+
+  if (processed_variables.count(overridden_name))
+    return true;
+
+  if (!stack_frame_sp)
+    return true;
+
+  CompilerType target_type;
+
+  // If we're not binding the generic types, we need to set the self type as an
+  // opaque pointer type. This is necessary because we don't bind the generic
+  // parameters, and we can't have a type with unbound generics in a non-generic
+  // function.
+  if (is_self && bind_generic_types == lldb::eDontBind) {
+    target_type = ast_context.GetBuiltinRawPointerType();
+  } else {
+    CompilerType var_type = SwiftExpressionParser::ResolveVariable(
+        variable_sp, stack_frame_sp, runtime, use_dynamic, bind_generic_types);
+
+    Status error;
+    target_type = ast_context.ImportType(var_type, error);
+  }
+
+  // If the import failed, give up.
+  if (!target_type.IsValid())
+    return true;
+
+  // If we couldn't fully realize the type, then we aren't going
+  // to get very far making a local out of it, so discard it here.
+  Log *log = GetLog(LLDBLog::Types | LLDBLog::Expressions);
+  if (!SwiftASTContext::IsFullyRealized(target_type)) {
+    if (log)
+      log->Printf("Discarding local %s because we couldn't fully realize it, "
+                  "our best attempt was: %s.",
+                  name_cstr,
+                  target_type.GetDisplayTypeName().AsCString("<unknown>"));
+    // Not realizing self is a fatal error for an expression and the
+    // Swift compiler error alone is not particularly useful.
+    if (is_self)
+      return false;
+    return true;
+  }
+
+  if (log && is_self)
+    if (swift::Type swift_type = GetSwiftType(target_type)) {
+      std::string s;
+      llvm::raw_string_ostream ss(s);
+      swift_type->dump(ss);
+      ss.flush();
+      log->Printf("Adding injected self: type (%p) context(%p) is: %s",
+                  static_cast<void *>(swift_type.getPointer()),
+                  static_cast<void *>(ast_context.GetASTContext()), s.c_str());
+    }
+  // A one-off clone of variable_sp with the type replaced by target_type.
+  auto patched_variable_sp = std::make_shared<lldb_private::Variable>(
+      0, variable_sp->GetName().GetCString(), "",
+      std::make_shared<lldb_private::SymbolFileType>(
+          *variable_sp->GetType()->GetSymbolFile(),
+          std::make_shared<lldb_private::Type>(
+              0, variable_sp->GetType()->GetSymbolFile(),
+              variable_sp->GetType()->GetName(), llvm::None,
+              variable_sp->GetType()->GetSymbolContextScope(), LLDB_INVALID_UID,
+              Type::eEncodingIsUID, variable_sp->GetType()->GetDeclaration(),
+              target_type, lldb_private::Type::ResolveState::Full,
+              variable_sp->GetType()->GetPayload())),
+      variable_sp->GetScope(), variable_sp->GetSymbolContextScope(),
+      variable_sp->GetScopeRange(),
+      const_cast<lldb_private::Declaration *>(&variable_sp->GetDeclaration()),
+      variable_sp->LocationExpression(), variable_sp->IsExternal(),
+      variable_sp->IsArtificial(),
+      variable_sp->GetLocationIsConstantValueData(),
+      variable_sp->IsStaticMember(), variable_sp->IsConstant());
+  SwiftASTManipulatorBase::VariableMetadataSP metadata_sp(
+      new SwiftASTManipulatorBase::VariableMetadataVariable(
+          patched_variable_sp));
+  SwiftASTManipulator::VariableInfo variable_info(
+      target_type, ast_context.GetASTContext()->getIdentifier(overridden_name),
+      metadata_sp,
+      variable_sp->IsConstant() ? swift::VarDecl::Introducer::Let
+                                : swift::VarDecl::Introducer::Var);
+
+  local_variables.push_back(variable_info);
+  processed_variables.insert(overridden_name);
+  return true;
+}
+
+ /// Create a \c VariableInfo record for each visible variable.
+ static bool RegisterAllVariables(
+     SymbolContext &sc, lldb::StackFrameSP &stack_frame_sp,
+     SwiftASTContextForExpressions &ast_context,
+     llvm::SmallVectorImpl<SwiftASTManipulator::VariableInfo> &local_variables,
+     lldb::DynamicValueType use_dynamic,
+     lldb::BindGenericTypes bind_generic_types) {
+   LLDB_SCOPED_TIMER();
+   if (!sc.block && !sc.function)
+     return true;
+
+   Block *block = sc.block;
+   Block *top_block = block->GetContainingInlinedBlock();
+
+   if (!top_block)
+     top_block = &sc.function->GetBlock(true);
+
+   SwiftLanguageRuntime *language_runtime = nullptr;
+
+   if (stack_frame_sp)
+     language_runtime =
+         SwiftLanguageRuntime::Get(stack_frame_sp->GetThread()->GetProcess());
+
+   // The module scoped variables are stored at the CompUnit level, so
+   // after we go through the current context, then we have to take one
+   // more pass through the variables in the CompUnit.
+   VariableList variables;
+
+   // Proceed from the innermost scope outwards, adding all variables
+   // not already shadowed by an inner declaration.
+   llvm::SmallDenseSet<const char *, 8> processed_names;
+   bool done = false;
+   do {
+     // Iterate over all parent contexts *including* the top_block.
+     if (block == top_block)
+       done = true;
+     bool can_create = true;
+     bool get_parent_variables = false;
+     bool stop_if_block_is_inlined_function = true;
+
+     block->AppendVariables(
+         can_create, get_parent_variables, stop_if_block_is_inlined_function,
+         [](Variable *) { return true; }, &variables);
+
+     if (!done)
+       block = block->GetParent();
+   } while (block && !done);
+
+   // Also add local copies of globals. This is in many cases redundant
+   // work because the globals would also be found in the expression
+   // context's Swift module, but it allows a limited form of
+   // expression evaluation to work even if the Swift module failed to
+   // load, as long as the module isn't necessary to resolve the type
+   // or aother symbols in the expression.
+   if (sc.comp_unit) {
+     lldb::VariableListSP globals_sp = sc.comp_unit->GetVariableList(true);
+     if (globals_sp)
+       variables.AddVariables(globals_sp.get());
+   }
+
+   for (size_t vi = 0, ve = variables.GetSize(); vi != ve; ++vi)
+     if (!AddVariableInfo({variables.GetVariableAtIndex(vi)}, stack_frame_sp,
+                          ast_context, language_runtime, processed_names,
+                          local_variables, use_dynamic, bind_generic_types))
+       return false;
+   return true;
+ }
+
 static SwiftPersistentExpressionState *
 GetPersistentState(Target *target, ExecutionContext &exe_ctx) {
   auto exe_scope = exe_ctx.GetBestExecutionContextScope();
@@ -279,14 +460,35 @@ SwiftUserExpression::GetTextAndSetExpressionParser(
   Log *log = GetLog(LLDBLog::Expressions);
   uint32_t first_body_line = 0;
 
+  lldb::TargetSP target_sp;
+  SymbolContext sc;
+  lldb::StackFrameSP stack_frame;
+  if (exe_scope) {
+    target_sp = exe_scope->CalculateTarget();
+    stack_frame = exe_scope->CalculateStackFrame();
+
+    if (stack_frame) {
+       sc = stack_frame->GetSymbolContext(lldb::eSymbolContextEverything);
+    }
+  }
+  llvm::SmallVector<SwiftASTManipulator::VariableInfo> local_variables;
+
+  if (!RegisterAllVariables(sc, stack_frame, *m_swift_ast_ctx, local_variables,
+                            m_options.GetUseDynamic(),
+                            m_options.GetBindGenericTypes())) {
+    diagnostic_manager.PutString(
+        eDiagnosticSeverityError,
+        "Couldn't realize Swift AST type of self. Hint: using `v` to "
+        "directly inspect variables and fields may still work.");
+    return ParseResult::retry_no_bind_generic_params;
+  }
+
   if (!source_code->GetText(m_transformed_text, m_options.GetLanguage(),
-                            m_needs_object_ptr,
-                            m_in_static_method,
-                            m_is_class,
-                            m_is_weak_self,  m_options, exe_ctx,
+                            m_needs_object_ptr, m_in_static_method, m_is_class,
+                            m_is_weak_self, m_options, exe_ctx,
                             first_body_line)) {
     diagnostic_manager.PutString(eDiagnosticSeverityError,
-                                  "couldn't construct expression body");
+                                 "couldn't construct expression body");
     return ParseResult::unrecoverable_error;
   }
 
@@ -303,7 +505,8 @@ SwiftUserExpression::GetTextAndSetExpressionParser(
     m_materializer_up.reset(new Materializer());
 
   auto *swift_parser =
-      new SwiftExpressionParser(exe_scope, *m_swift_ast_ctx, *this, m_options);
+      new SwiftExpressionParser(exe_scope, *m_swift_ast_ctx, *this,
+                                std::move(local_variables), m_options);
   SwiftExpressionParser::ParseResult parse_result =
       swift_parser->Parse(diagnostic_manager, first_body_line,
                           first_body_line + source_code->GetNumBodyLines());
